+++ src/utils/writer_utils.rs
@@ -6,644 +6,218 @@
 //! writer utilities.
 
 
-use errors;
-use quick_xml::{Element, XmlWriter};
-use quick_xml::Event::*;
-use channels::{Channel, Item};
-use std::io::Cursor;
+
+use channels::{Category, Channel, Cloud, Enclosure, Image, Item, Guid, Source, TextInput};
+use rss;
+use utils::string_utils;
 
 
 /// Construct xml from a `Channel`.
 pub fn write(channel: Channel) -> Vec<u8> {
-    let mut writer = XmlWriter::new(Cursor::new(Vec::new()));
-
-    let xml_tag_str = "?xml";
-    let mut xml_tag = Element::new(xml_tag_str);
-    xml_tag.push_attribute(b"version", "1.0");
-    xml_tag.push_attribute(b"encoding", "UTF-8");
-    writer.write(Start(xml_tag)).expect(errors::tag_start_error(xml_tag_str).as_str());
-    writer.write(End(Element::new(xml_tag_str)))
-        .expect(errors::tag_end_error(xml_tag_str).as_str());
-
-    let rss_tag_str = "channels";
-    let mut rss_tag = Element::new(rss_tag_str);
-    rss_tag.push_attribute(b"xmlns:atom", "http://www.w3.org/2005/Atom");
-    rss_tag.push_attribute(b"xmlns:itunes",
-                           "http://www.itunes.com/dtds/podcast-1.0.dtd");
-    rss_tag.push_attribute(b"version", "2.0");
-    writer.write(Start(rss_tag)).expect(errors::tag_start_error(rss_tag_str).as_str());
-
-    write_channel(writer.clone(), channel.clone());
-
-    writer.write(End(Element::new(rss_tag_str)))
-        .expect(errors::tag_end_error(rss_tag_str).as_str());
-    writer.into_inner().into_inner()
+    let rss_channel = convert_channel(channel);
+    let s = rss_channel.to_string();
+    s.into_bytes()
 }
 
 
-fn write_channel(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    let channel_tag_str = "channels";
-    let channel_tag = Element::new(channel_tag_str);
-    writer.write(Start(channel_tag))
-        .expect(errors::tag_start_error(channel_tag_str).as_str());
-
-    write_title(writer.clone(), channel.clone());
-    write_link(writer.clone(), channel.clone());
-    write_description(writer.clone(), channel.clone());
-    write_generator(writer.clone(), channel.clone());
-    write_docs(writer.clone(), channel.clone());
-    write_language(writer.clone(), channel.clone());
-    write_copyright(writer.clone(), channel.clone());
-    write_managing_editor(writer.clone(), channel.clone());
-    write_web_master(writer.clone(), channel.clone());
-    write_cloud(writer.clone(), channel.clone());
-    write_ttl(writer.clone(), channel.clone());
-    write_categories(writer.clone(), channel.clone());
-    write_rating(writer.clone(), channel.clone());
-    write_pub_date(writer.clone(), channel.clone());
-    write_last_build_date(writer.clone(), channel.clone());
-    write_image(writer.clone(), channel.clone());
-    write_text_input(writer.clone(), channel.clone());
-    write_skip_hours(writer.clone(), channel.clone());
-    write_skip_days(writer.clone(), channel.clone());
-    write_items(writer.clone(), channel.clone());
-
-    writer.write(End(Element::new(channel_tag_str)))
-        .expect(errors::tag_end_error(channel_tag_str).as_str());
-}
-
-
-fn write_title(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    let title_tag_str = "title";
-    let title_tag = Element::new(title_tag_str);
-    writer.write(Start(title_tag)).expect(errors::tag_start_error(title_tag_str).as_str());
-    writer.write(Text(Element::new(channel.title().as_str())))
-        .expect(errors::tag_text_error(title_tag_str).as_str());
-    writer.write(End(Element::new(title_tag_str)))
-        .expect(errors::tag_end_error(title_tag_str).as_str());
-}
-
-
-fn write_link(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    let link_tag_str = "link";
-    let link_tag = Element::new(link_tag_str);
-    writer.write(Start(link_tag)).expect(errors::tag_start_error(link_tag_str).as_str());
-    writer.write(Text(Element::new(channel.link().as_str())))
-        .expect(errors::tag_text_error(link_tag_str).as_str());
-    writer.write(End(Element::new(link_tag_str)))
-        .expect(errors::tag_end_error(link_tag_str).as_str());
-}
-
-
-fn write_description(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    let description_tag_str = "description";
-    let description_tag = Element::new(description_tag_str);
-    writer.write(Start(description_tag))
-        .expect(errors::tag_start_error(description_tag_str).as_str());
-    writer.write(Text(Element::new(channel.description().as_str())))
-        .expect(errors::tag_text_error(description_tag_str).as_str());
-    writer.write(End(Element::new(description_tag_str)))
-        .expect(errors::tag_end_error(description_tag_str).as_str());
-}
-
-
-fn write_generator(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.generator().is_some() {
-        let generator_tag_str = "generator";
-        let generator_tag = Element::new(generator_tag_str);
-        writer.write(Start(generator_tag))
-            .expect(errors::tag_start_error(generator_tag_str).as_str());
-        writer.write(Text(Element::new(channel.generator().unwrap().as_str())))
-            .expect(errors::tag_text_error(generator_tag_str).as_str());
-        writer.write(End(Element::new(generator_tag_str)))
-            .expect(errors::tag_end_error(generator_tag_str).as_str());
+// convert rss channel from feed channel
+fn convert_channel(channel: Channel) -> rss::Channel {
+    rss::Channel {
+        title: channel.title(),
+        link: channel.link(),
+        description: channel.description(),
+        language: channel.language(),
+        copyright: channel.copyright(),
+        managing_editor: channel.managing_editor(),
+        webmaster: channel.web_master(),
+        pub_date: string_utils::option_date_to_option_string(channel.pub_date()),
+        last_build_date: string_utils::option_date_to_option_string(channel.last_build_date()),
+        categories: convert_categories(channel.categories()),
+        generator: channel.generator(),
+        docs: channel.docs(),
+        cloud: convert_cloud(channel.cloud()),
+        ttl: string_utils::option_i64_to_option_string(channel.ttl()),
+        image: convert_image(channel.image()),
+        // rating: channel.rating(),
+        text_input: convert_text_input(channel.text_input()),
+        skip_hours: convert_skip_hours(channel.skip_hours()),
+        skip_days: convert_skip_days(channel.skip_days()),
+        items: convert_items(channel.items()),
+        itunes_ext: None,
+        dublin_core_ext: None,
+        ..Default::default()
     }
 }
 
 
-fn write_docs(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.docs().is_some() {
-        let docs_tag_str = "docs";
-        let docs_tag = Element::new(docs_tag_str);
-        writer.write(Start(docs_tag))
-            .expect(errors::tag_start_error(docs_tag_str).as_str());
-        writer.write(Text(Element::new(channel.docs().unwrap().as_str())))
-            .expect(errors::tag_text_error(docs_tag_str).as_str());
-        writer.write(End(Element::new(docs_tag_str)))
-            .expect(errors::tag_end_error(docs_tag_str).as_str());
-    }
-}
-
-
-fn write_language(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.language().is_some() {
-        let language_tag_str = "language";
-        let language_tag = Element::new(language_tag_str);
-        writer.write(Start(language_tag))
-            .expect(errors::tag_start_error(language_tag_str).as_str());
-        writer.write(Text(Element::new(channel.language().unwrap().as_str())))
-            .expect(errors::tag_text_error(language_tag_str).as_str());
-        writer.write(End(Element::new(language_tag_str)))
-            .expect(errors::tag_end_error(language_tag_str).as_str());
-    }
-}
-
-
-fn write_copyright(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.copyright().is_some() {
-        let copyright_tag_str = "copyright";
-        let copyright_tag = Element::new(copyright_tag_str);
-        writer.write(Start(copyright_tag))
-            .expect(errors::tag_start_error(copyright_tag_str).as_str());
-        writer.write(Text(Element::new(channel.copyright().unwrap().as_str())))
-            .expect(errors::tag_text_error(copyright_tag_str).as_str());
-        writer.write(End(Element::new(copyright_tag_str)))
-            .expect(errors::tag_end_error(copyright_tag_str).as_str());
-    }
-}
-
-
-fn write_managing_editor(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.managing_editor().is_some() {
-        let managing_editor_tag_str = "managingEditor";
-        let managing_editor_tag = Element::new(managing_editor_tag_str);
-        writer.write(Start(managing_editor_tag))
-            .expect(errors::tag_start_error(managing_editor_tag_str).as_str());
-        writer.write(Text(Element::new(channel.managing_editor().unwrap().as_str())))
-            .expect(errors::tag_text_error(managing_editor_tag_str).as_str());
-        writer.write(End(Element::new(managing_editor_tag_str)))
-            .expect(errors::tag_end_error(managing_editor_tag_str).as_str());
-    }
-}
-
-
-fn write_web_master(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.web_master().is_some() {
-        let web_master_tag_str = "webMaster";
-        let web_master_tag = Element::new(web_master_tag_str);
-        writer.write(Start(web_master_tag))
-            .expect(errors::tag_start_error(web_master_tag_str).as_str());
-        writer.write(Text(Element::new(channel.web_master().unwrap().as_str())))
-            .expect(errors::tag_text_error(web_master_tag_str).as_str());
-        writer.write(End(Element::new(web_master_tag_str)))
-            .expect(errors::tag_end_error(web_master_tag_str).as_str());
-    }
-}
-
-
-fn write_cloud(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.cloud().is_some() {
-        let cloud_tag_str = "cloud";
-        let mut cloud_tag = Element::new(cloud_tag_str);
-        cloud_tag.push_attribute(b"domain", channel.cloud().unwrap().domain().as_str());
-        cloud_tag.push_attribute(b"port",
-                                 channel.cloud().unwrap().port().to_string().as_str());
-        cloud_tag.push_attribute(b"path", channel.cloud().unwrap().path().as_str());
-        cloud_tag.push_attribute(b"registerProcedure",
-                                 channel.cloud().unwrap().register_procedure().as_str());
-        cloud_tag.push_attribute(b"protocol", channel.cloud().unwrap().protocol().as_str());
-        writer.write(Start(cloud_tag))
-            .expect(errors::tag_start_error(cloud_tag_str).as_str());
-        writer.write(End(Element::new(cloud_tag_str)))
-            .expect(errors::tag_end_error(cloud_tag_str).as_str());
-    }
-}
-
-
-fn write_ttl(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.ttl().is_some() {
-        let ttl_tag_str = "ttl";
-        let ttl_tag = Element::new(ttl_tag_str);
-        writer.write(Start(ttl_tag))
-            .expect(errors::tag_start_error(ttl_tag_str).as_str());
-        writer.write(Text(Element::new(channel.ttl().unwrap().to_string().as_str())))
-            .expect(errors::tag_text_error(ttl_tag_str).as_str());
-        writer.write(End(Element::new(ttl_tag_str)))
-            .expect(errors::tag_end_error(ttl_tag_str).as_str());
-    }
-}
-
-
-fn write_categories(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.categories().is_some() {
-        for category in channel.categories().unwrap() {
-            let category_tag_str = "category";
-            let mut category_tag = Element::new(category_tag_str);
-            if category.domain().is_some() {
-                category_tag.push_attribute(b"domain", category.domain().unwrap().as_str());
-            }
-            writer.write(Start(category_tag))
-                .expect(errors::tag_start_error(category_tag_str).as_str());
-            writer.write(Text(Element::new(category.category())))
-                .expect(errors::tag_text_error(category_tag_str).as_str());
-            writer.write(End(Element::new(category_tag_str)))
-                .expect(errors::tag_end_error(category_tag_str).as_str());
+// Convert rss categories from feed categories
+fn convert_categories(cats_opt: Option<Vec<Category>>) -> Vec<rss::Category> {
+    let mut rss_cats = Vec::new();
+    if cats_opt.is_some() {
+        let cats = cats_opt.unwrap();
+        for cat in cats {
+            let rss_cat = rss::Category {
+                name: cat.name(),
+                domain: cat.domain(),
+            };
+            rss_cats.push(rss_cat);
         }
     }
+    rss_cats
 }
 
 
-fn write_rating(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.rating().is_some() {
-        let rating_tag_str = "rating";
-        let rating_tag = Element::new(rating_tag_str);
-        writer.write(Start(rating_tag))
-            .expect(errors::tag_start_error(rating_tag_str).as_str());
-        writer.write(Text(Element::new(channel.rating().unwrap().as_str())))
-            .expect(errors::tag_text_error(rating_tag_str).as_str());
-        writer.write(End(Element::new(rating_tag_str)))
-            .expect(errors::tag_end_error(rating_tag_str).as_str());
+// Convert rss channel cloud from feed channel cloud
+fn convert_cloud(cloud_opt: Option<Cloud>) -> Option<rss::Cloud> {
+    if cloud_opt.is_none() {
+        None
+    } else {
+        let cloud = cloud_opt.unwrap();
+        let rss_cloud = rss::Cloud {
+            domain: cloud.domain(),
+            port: cloud.port().to_string(),
+            path: cloud.path(),
+            register_procedure: cloud.register_procedure(),
+            protocol: cloud.protocol(),
+        };
+        Some(rss_cloud)
     }
 }
 
 
-fn write_pub_date(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.pub_date().is_some() {
-        let pub_date_tag_str = "pubDate";
-        let pub_date_tag = Element::new(pub_date_tag_str);
-        writer.write(Start(pub_date_tag))
-            .expect(errors::tag_start_error(pub_date_tag_str).as_str());
-        writer.write(Text(Element::new(channel.pub_date().unwrap().to_rfc2822())))
-            .expect(errors::tag_text_error(pub_date_tag_str).as_str());
-        writer.write(End(Element::new(pub_date_tag_str)))
-            .expect(errors::tag_end_error(pub_date_tag_str).as_str());
+// Convert rss channel image from feed channel image
+fn convert_image(image_opt: Option<Image>) -> Option<rss::Image> {
+    if image_opt.is_none() {
+        None
+    } else {
+        let image = image_opt.unwrap();
+        let rss_image = rss::Image {
+            url: image.url(),
+            title: image.title(),
+            link: image.link(),
+            width: string_utils::i64_to_option_string(image.width()),
+            height: string_utils::i64_to_option_string(image.height()),
+            description: image.description(),
+        };
+        Some(rss_image)
     }
 }
 
 
-fn write_last_build_date(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.last_build_date().is_some() {
-        let last_build_date_tag_str = "lastBuildDate";
-        let last_build_date_tag = Element::new(last_build_date_tag_str);
-        writer.write(Start(last_build_date_tag))
-            .expect(errors::tag_start_error(last_build_date_tag_str).as_str());
-        writer.write(Text(Element::new(channel.last_build_date().unwrap().to_rfc2822())))
-            .expect(errors::tag_text_error(last_build_date_tag_str).as_str());
-        writer.write(End(Element::new(last_build_date_tag_str)))
-            .expect(errors::tag_end_error(last_build_date_tag_str).as_str());
+// Convert rss channel text input from feed channel text input
+fn convert_text_input(text_input_opt: Option<TextInput>) -> Option<rss::TextInput> {
+    if text_input_opt.is_none() {
+        None
+    } else {
+        let text_input = text_input_opt.unwrap();
+        let rss_text_input = rss::TextInput {
+            title: text_input.title(),
+            description: text_input.description(),
+            name: text_input.name(),
+            link: text_input.link(),
+        };
+        Some(rss_text_input)
     }
 }
 
 
-fn write_image(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.image().is_some() {
-        let image_tag_str = "image";
-        let image_tag = Element::new(image_tag_str);
-        writer.write(Start(image_tag))
-            .expect(errors::tag_start_error(image_tag_str).as_str());
-
-        let image_link_tag_str = "link";
-        let image_link_tag = Element::new(image_link_tag_str);
-        writer.write(Start(image_link_tag))
-            .expect(errors::tag_start_error(image_link_tag_str).as_str());
-        writer.write(Text(Element::new(channel.image()
-                .unwrap()
-                .link()
-                .as_str())))
-            .expect(errors::tag_text_error(image_link_tag_str).as_str());
-        writer.write(End(Element::new(image_link_tag_str)))
-            .expect(errors::tag_end_error(image_link_tag_str).as_str());
-
-        let image_url_tag_str = "url";
-        let image_url_tag = Element::new(image_url_tag_str);
-        writer.write(Start(image_url_tag))
-            .expect(errors::tag_start_error(image_url_tag_str).as_str());
-        writer.write(Text(Element::new(channel.image()
-                .unwrap()
-                .url()
-                .as_str())))
-            .expect(errors::tag_text_error(image_url_tag_str).as_str());
-        writer.write(End(Element::new(image_url_tag_str)))
-            .expect(errors::tag_end_error(image_url_tag_str).as_str());
-
-        let image_title_tag_str = "title";
-        let image_title_tag = Element::new(image_title_tag_str);
-        writer.write(Start(image_title_tag))
-            .expect(errors::tag_start_error(image_title_tag_str).as_str());
-        writer.write(Text(Element::new(channel.image()
-                .unwrap()
-                .title()
-                .as_str())))
-            .expect(errors::tag_text_error(image_title_tag_str).as_str());
-        writer.write(End(Element::new(image_title_tag_str)))
-            .expect(errors::tag_end_error(image_title_tag_str).as_str());
-
-        let image_width_tag_str = "width";
-        let image_width_tag = Element::new(image_width_tag_str);
-        writer.write(Start(image_width_tag))
-            .expect(errors::tag_start_error(image_width_tag_str).as_str());
-        writer.write(Text(Element::new(channel.image()
-                .unwrap()
-                .width()
-                .to_string()
-                .as_str())))
-            .expect(errors::tag_text_error(image_width_tag_str).as_str());
-        writer.write(End(Element::new(image_width_tag_str)))
-            .expect(errors::tag_end_error(image_width_tag_str).as_str());
-
-        let image_height_tag_str = "height";
-        let image_height_tag = Element::new(image_height_tag_str);
-        writer.write(Start(image_height_tag))
-            .expect(errors::tag_start_error(image_height_tag_str).as_str());
-        writer.write(Text(Element::new(channel.image()
-                .unwrap()
-                .height()
-                .to_string()
-                .as_str())))
-            .expect(errors::tag_text_error(image_height_tag_str).as_str());
-        writer.write(End(Element::new(image_height_tag_str)))
-            .expect(errors::tag_end_error(image_height_tag_str).as_str());
-
-        if channel.image().unwrap().description().is_some() {
-            let image_description_tag_str = "description";
-            let image_description_tag = Element::new(image_description_tag_str);
-            writer.write(Start(image_description_tag))
-                .expect(errors::tag_start_error(image_description_tag_str).as_str());
-            writer.write(Text(Element::new(channel.image()
-                    .unwrap()
-                    .description()
-                    .unwrap()
-                    .as_str())))
-                .expect(errors::tag_text_error(image_description_tag_str).as_str());
-            writer.write(End(Element::new(image_description_tag_str)))
-                .expect(errors::tag_end_error(image_description_tag_str).as_str());
-        }
-
-        writer.write(End(Element::new(image_tag_str)))
-            .expect(errors::tag_end_error(image_tag_str).as_str());
-    }
-}
-
-fn write_text_input(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.text_input().is_some() {
-        let text_input_str = "textInput";
-        let text_input = Element::new(text_input_str);
-        writer.write(Start(text_input))
-            .expect(errors::tag_start_error(text_input_str).as_str());
-
-        let title_tag_str = "title";
-        let title_tag = Element::new(title_tag_str);
-        writer.write(Start(title_tag))
-            .expect(errors::tag_start_error(title_tag_str).as_str());
-        writer.write(Text(Element::new(channel.text_input().unwrap().title().as_str())))
-            .expect(errors::tag_text_error(title_tag_str).as_str());
-        writer.write(End(Element::new(title_tag_str)))
-            .expect(errors::tag_end_error(title_tag_str).as_str());
-
-        let description_tag_str = "description";
-        let description_tag = Element::new(description_tag_str);
-        writer.write(Start(description_tag))
-            .expect(errors::tag_start_error(description_tag_str).as_str());
-        writer.write(Text(Element::new(channel.description().as_str())))
-            .expect(errors::tag_text_error(description_tag_str).as_str());
-        writer.write(End(Element::new(description_tag_str)))
-            .expect(errors::tag_end_error(description_tag_str).as_str());
-
-        let name_tag_str = "name";
-        let name_tag = Element::new(name_tag_str);
-        writer.write(Start(name_tag))
-            .expect(errors::tag_start_error(name_tag_str).as_str());
-        writer.write(Text(Element::new(channel.text_input().unwrap().name().as_str())))
-            .expect(errors::tag_text_error(name_tag_str).as_str());
-        writer.write(End(Element::new(name_tag_str)))
-            .expect(errors::tag_end_error(name_tag_str).as_str());
-
-        let link_tag_str = "link";
-        let link_tag = Element::new(link_tag_str);
-        writer.write(Start(link_tag))
-            .expect(errors::tag_start_error(link_tag_str).as_str());
-        writer.write(Text(Element::new(channel.text_input().unwrap().link().as_str())))
-            .expect(errors::tag_text_error(link_tag_str).as_str());
-        writer.write(End(Element::new(link_tag_str)))
-            .expect(errors::tag_end_error(link_tag_str).as_str());
-
-        writer.write(End(Element::new(text_input_str)))
-            .expect(errors::tag_end_error(text_input_str).as_str());
-    }
-}
-
-
-fn write_skip_hours(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.skip_hours().is_some() {
-        let skip_hours_tag_str = "skipHours";
-        let skip_hours_tag = Element::new(skip_hours_tag_str);
-        writer.write(Start(skip_hours_tag))
-            .expect(errors::tag_start_error(skip_hours_tag_str).as_str());
-        for hour in channel.skip_hours().unwrap() {
-            let hour_tag_str = "hour";
-            let hour_tag = Element::new(hour_tag_str);
-            writer.write(Start(hour_tag))
-                .expect(errors::tag_start_error(hour_tag_str).as_str());
-            writer.write(Text(Element::new(hour.to_string().as_str())))
-                .expect(errors::tag_text_error(hour_tag_str).as_str());
-            writer.write(End(Element::new(hour_tag_str)))
-                .expect(errors::tag_end_error(hour_tag_str).as_str());
-        }
-        writer.write(End(Element::new(skip_hours_tag_str)))
-            .expect(errors::tag_end_error(skip_hours_tag_str).as_str());
-    }
-}
-
-
-fn write_skip_days(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.skip_days().is_some() {
-        let skip_days_tag_str = "skipDays";
-        let skip_days_tag = Element::new(skip_days_tag_str);
-        writer.write(Start(skip_days_tag))
-            .expect(errors::tag_start_error(skip_days_tag_str).as_str());
-        for day in channel.skip_days().unwrap() {
-            let day_tag_str = "day";
-            let day_tag = Element::new(day_tag_str);
-            writer.write(Start(day_tag))
-                .expect(errors::tag_start_error(day_tag_str).as_str());
-            writer.write(Text(Element::new(day.as_str())))
-                .expect(errors::tag_text_error(day_tag_str).as_str());
-            writer.write(End(Element::new(day_tag_str)))
-                .expect(errors::tag_end_error(day_tag_str).as_str());
-        }
-        writer.write(End(Element::new(skip_days_tag_str)))
-            .expect(errors::tag_end_error(skip_days_tag_str).as_str());
-    }
-}
-
-
-fn write_items(mut writer: XmlWriter<Cursor<Vec<u8>>>, channel: Channel) {
-    if channel.items().is_some() {
-        for item in channel.items().unwrap() {
-            let item_tag_str = "item";
-            let item_tag = Element::new(item_tag_str);
-            writer.write(Start(item_tag))
-                .expect(errors::tag_start_error(item_tag_str).as_str());
-
-            write_item_title(writer.clone(), item.clone());
-            write_item_link(writer.clone(), item.clone());
-            write_item_description(writer.clone(), item.clone());
-            write_item_author(writer.clone(), item.clone());
-            write_item_categories(writer.clone(), item.clone());
-            write_item_comments(writer.clone(), item.clone());
-            write_item_enclosure(writer.clone(), item.clone());
-            write_item_guid(writer.clone(), item.clone());
-            write_item_pub_date(writer.clone(), item.clone());
-            write_item_source(writer.clone(), item.clone());
-
-            writer.write(End(Element::new(item_tag_str)))
-                .expect(errors::tag_end_error(item_tag_str).as_str());
+// Convert rss channel skip hours from feed channel skip hours
+fn convert_skip_hours(skip_hours: Option<Vec<i64>>) -> Vec<String> {
+    let mut rss_skip_hours = Vec::new();
+    if skip_hours.is_some() {
+        let hours = skip_hours.unwrap();
+        for hour in hours {
+            rss_skip_hours.push(hour.to_string());
         }
     }
+    rss_skip_hours
 }
 
 
-fn write_item_title(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.title().is_some() {
-        let item_title_tag_str = "title";
-        let item_title_tag = Element::new(item_title_tag_str);
-        writer.write(Start(item_title_tag))
-            .expect(errors::tag_start_error(item_title_tag_str).as_str());
-        writer.write(Text(Element::new(item.title().unwrap().as_str())))
-            .expect(errors::tag_text_error(item_title_tag_str).as_str());
-        writer.write(End(Element::new(item_title_tag_str)))
-            .expect(errors::tag_end_error(item_title_tag_str).as_str());
-    }
-}
 
-
-fn write_item_link(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.link().is_some() {
-        let item_link_tag_str = "link";
-        let item_link_tag = Element::new(item_link_tag_str);
-        writer.write(Start(item_link_tag))
-            .expect(errors::tag_start_error(item_link_tag_str).as_str());
-        writer.write(Text(Element::new(item.link().unwrap().as_str())))
-            .expect(errors::tag_text_error(item_link_tag_str).as_str());
-        writer.write(End(Element::new(item_link_tag_str)))
-            .expect(errors::tag_end_error(item_link_tag_str).as_str());
-    }
-}
-
-
-fn write_item_description(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.description().is_some() {
-        let item_description_tag_str = "description";
-        let item_description_tag = Element::new(item_description_tag_str);
-        writer.write(Start(item_description_tag))
-            .expect(errors::tag_start_error(item_description_tag_str).as_str());
-        writer.write(Text(Element::new(item.description().unwrap().as_str())))
-            .expect(errors::tag_text_error(item_description_tag_str).as_str());
-        writer.write(End(Element::new(item_description_tag_str)))
-            .expect(errors::tag_end_error(item_description_tag_str).as_str());
-    }
-}
-
-
-fn write_item_author(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.author().is_some() {
-        let item_author_tag_str = "author";
-        let item_author_tag = Element::new(item_author_tag_str);
-        writer.write(Start(item_author_tag))
-            .expect(errors::tag_start_error(item_author_tag_str).as_str());
-        writer.write(Text(Element::new(item.author().unwrap().as_str())))
-            .expect(errors::tag_text_error(item_author_tag_str).as_str());
-        writer.write(End(Element::new(item_author_tag_str)))
-            .expect(errors::tag_end_error(item_author_tag_str).as_str());
-    }
-}
-
-
-fn write_item_categories(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.categories().is_some() {
-        for category in item.categories().unwrap() {
-            let item_category_tag_str = "category";
-            let mut item_category_tag = Element::new(item_category_tag_str);
-            if category.domain().is_some() {
-                item_category_tag.push_attribute(b"domain", category.domain().unwrap().as_str());
-            }
-            writer.write(Start(item_category_tag))
-                .expect(errors::tag_start_error(item_category_tag_str).as_str());
-            writer.write(Text(Element::new(category.category())))
-                .expect(errors::tag_text_error(item_category_tag_str).as_str());
-            writer.write(End(Element::new(item_category_tag_str)))
-                .expect(errors::tag_end_error(item_category_tag_str).as_str());
+// Convert rss channel skip days from feed channel skip days
+fn convert_skip_days(skip_days: Option<Vec<String>>) -> Vec<String> {
+    let mut rss_skip_days = Vec::new();
+    if skip_days.is_some() {
+        let days = skip_days.unwrap();
+        for day in days {
+            rss_skip_days.push(day);
         }
     }
+    rss_skip_days
 }
 
 
-fn write_item_comments(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.comments().is_some() {
-        let item_comments_tag_str = "comments";
-        let item_comments_tag = Element::new(item_comments_tag_str);
-        writer.write(Start(item_comments_tag))
-            .expect(errors::tag_start_error(item_comments_tag_str).as_str());
-        writer.write(Text(Element::new(item.comments().unwrap().as_str())))
-            .expect(errors::tag_text_error(item_comments_tag_str).as_str());
-        writer.write(End(Element::new(item_comments_tag_str)))
-            .expect(errors::tag_end_error(item_comments_tag_str).as_str());
-    }
-}
-
-
-fn write_item_enclosure(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.enclosure().is_some() {
-        let item_enclosure_tag_str = "enclosure";
-        let mut item_enclosure_tag = Element::new(item_enclosure_tag_str);
-        item_enclosure_tag.push_attribute(b"url", item.enclosure().unwrap().url().as_str());
-        item_enclosure_tag.push_attribute(b"length",
-                                          item.enclosure().unwrap().length().to_string().as_str());
-        item_enclosure_tag.push_attribute(b"type",
-                                          item.enclosure()
-                                              .unwrap()
-                                              .mime_type()
-                                              .as_str());
-        writer.write(Start(item_enclosure_tag))
-            .expect(errors::tag_start_error(item_enclosure_tag_str).as_str());
-
-    }
-}
-
-
-fn write_item_guid(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.guid().is_some() {
-        let item_guid_tag_str = "guid";
-        let mut item_guid_tag = Element::new(item_guid_tag_str);
-        item_guid_tag.push_attribute(b"isPermaLink",
-                                     item.guid().unwrap().is_permalink().to_string().as_str());
-        writer.write(Start(item_guid_tag))
-            .expect(errors::tag_start_error(item_guid_tag_str).as_str());
-        writer.write(Text(Element::new(item.guid().unwrap().value().as_str())))
-            .expect(errors::tag_text_error(item_guid_tag_str).as_str());
-        writer.write(End(Element::new(item_guid_tag_str)))
-            .expect(errors::tag_end_error(item_guid_tag_str).as_str());
-    }
-}
-
-
-fn write_item_pub_date(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.pub_date().is_some() {
-        let item_pub_date_tag_str = "pubDate";
-        let item_pub_date_tag = Element::new(item_pub_date_tag_str);
-        writer.write(Start(item_pub_date_tag))
-            .expect(errors::tag_start_error(item_pub_date_tag_str).as_str());
-        writer.write(Text(Element::new(item.pub_date().unwrap().to_rfc2822())))
-            .expect(errors::tag_text_error(item_pub_date_tag_str).as_str());
-        writer.write(End(Element::new(item_pub_date_tag_str)))
-            .expect(errors::tag_end_error(item_pub_date_tag_str).as_str());
-    }
-}
-
-
-fn write_item_source(mut writer: XmlWriter<Cursor<Vec<u8>>>, item: Item) {
-    if item.source().is_some() {
-        let item_source_tag_str = "source";
-        let mut item_source_tag = Element::new(item_source_tag_str);
-        item_source_tag.push_attribute(b"url", item.source().unwrap().url().to_string().as_str());
-        writer.write(Start(item_source_tag))
-            .expect(errors::tag_start_error(item_source_tag_str).as_str());
-        if item.source().unwrap().title().is_some() {
-            let src = item.source().unwrap().title().unwrap();
-            writer.write(Text(Element::new(src.as_str())))
-                .expect(errors::tag_text_error(item_source_tag_str).as_str());
+// Convert rss channel items from feed channel items
+fn convert_items(items_opt: Option<Vec<Item>>) -> Vec<rss::Item> {
+    let mut rss_items = Vec::new();
+    if items_opt.is_some() {
+        let items = items_opt.unwrap();
+        for item in items {
+            let rss_item = rss::Item {
+                title: item.title(),
+                link: item.link(),
+                description: item.description(),
+                author: item.author(),
+                categories: convert_categories(item.categories()),
+                comments: item.comments(),
+                enclosure: convert_enclosure(item.enclosure()),
+                guid: convert_guid(item.guid()),
+                pub_date: string_utils::option_date_to_option_string(item.pub_date()),
+                source: convert_source(item.source()),
+                content: None,
+                itunes_ext: None,
+                dublin_core_ext: None,
+                ..Default::default()
+            };
+            rss_items.push(rss_item);
         }
-        writer.write(End(Element::new(item_source_tag_str)))
-            .expect(errors::tag_end_error(item_source_tag_str).as_str());
+    }
+    rss_items
+}
+
+
+// Convert rss item enclosure from feed item enclosure
+fn convert_enclosure(enc_opt: Option<Enclosure>) -> Option<rss::Enclosure> {
+    if enc_opt.is_none() {
+        None
+    } else {
+        let enc = enc_opt.unwrap();
+        let rss_enclosure = rss::Enclosure {
+            url: enc.url(),
+            length: enc.length().to_string(),
+            mime_type: enc.mime_type(),
+        };
+        Some(rss_enclosure)
+    }
+}
+
+
+// Convert rss item guid from feed item guid
+fn convert_guid(guid_opt: Option<Guid>) -> Option<rss::Guid> {
+    if guid_opt.is_none() {
+        None
+    } else {
+        let guid = guid_opt.unwrap();
+        let rss_guid = rss::Guid {
+            value: guid.value(),
+            is_permalink: guid.permalink(),
+        };
+        Some(rss_guid)
+    }
+}
+
+
+// Convert rss item source from feed item source
+fn convert_source(src_opt: Option<Source>) -> Option<rss::Source> {
+    if src_opt.is_none() {
+        None
+    } else {
+        let src = src_opt.unwrap();
+        let rss_src = rss::Source {
+            url: src.url(),
+            title: src.title(),
+        };
+        Some(rss_src)
     }
 }